Removed update of task state from application submission GET - moved to prior POST method on the travel screen
